Add files via upload
Nowadays, for humans, herb plants are important, as well as for the medical field. In
this paper, for plant leaf disease detection, we are using four plants, such as chilies, cotton,
rice, and tomatoes. For the purpose of analysis in that paper, for the classification of whether
the leaf is unhealthy or healthy, Phyllanthus Elegans Wall is used. The function of this plant
can cure breast cancer, so we chose it. For breast cancer patients, there is a need for rather
alternative methods such as the use of technology in medicine, hospitalization, surgery, and
chemotherapy. In the agriculture field, using technology to identify the quality of leaf from
classification, image processing, and image acquisition, for the analysis quality of leaves. For
the color transformation structure using HSV to input RCB image, the most important thing
for this part is the segmentation for the method of image processing. On the basis of shape
and color, the analysis of leaf disease images is applied. Finally, as the classification method,
employ with a back-propagation algorithm a feed-forward neural network. The result
compares Radial Basis Function (RBF) and Multi-layer Perceptron (MLP), with the difference
displayed in percentage of accuracy. For neural networks, RBF and MLP are algorithms. We
give better accuracy and performance using the classifier of the neural network